ui_backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.20
aiofiles==23.2.1
slowapi==0.1.8